+++ ChangeLog.md
@@ -11,6 +11,7 @@
 - Some enemies's poise now scales with party size
 - New Challenge: Placebos - Items have a chance to do nothing
 - Tweaked room rarities, largely to make Battles and Events more common
+- New Item: Panacea - Cures up to 2 debuffs on the user
 ## Prophets of the Labyrinth Version 0.12.0:
 - Reintroduced Light and Darkness elements
 - Reworked Martial Artist starting gear